Refactored page names
