+++ lib/pages/payment_method_detail.dart
